Restructured Plant Animation Scripts
Added:
Latitude - ScriptableObject to hold latitude methods and data if we add some

ILatitudeDependant - interface for things that change depending on the latitude, like plant animations. Just has ReturnLatitude()

Wind - ScriptableObject with wind methods and 2 values to play with in the inspector - Gust Frequency and Gust Strength Loop Frequency. Basically how fast the wooshes woosh and how fast the wooshes alternate from big woosh to lil woosh.

IWindReactive - Interface for stuff that reacts to wind. Has Gust() and GustStrength()

Sun - Scriptableobject with data and methods. SunZenith can eventually be like where the sun is overhead exactly, but for now it's a wave like the wind. If we tweak it it will actually match an orbiting sun though!

ISunReactive - has SunZenith()

PlantAnimator - This inherits all the new interfaces. They each update the animator on the gameobject. I collected all the methods in AnimUpdate().

PlantManager - I tried to add a method that would access all the plantanimators on all the plant gameobjects and do AnimUpdate() on the whole list in void update, but something wasn't working. Either I wasn't getting the PlantAnimator component or the PlantAnimator wasn't able to get the AnimatorController component. Either way, the animator is null which is causing errors.